feat : get category, subcategory, item list refactor : Item entity category -> order
