+++ backend/src/main/java/shop/ninescent/mall/category/repository/GetItemListRepository.java
@@ -1,7 +1,7 @@
 package shop.ninescent.mall.category.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.ninescent.mall.category.domain.Item;
+import shop.ninescent.mall.order.domain.Item;
 
 import java.util.List;
